NVSHAS-6166: k8s platform reported version is not aligned with actual version
