+++ controller/controller.go
@@ -289,7 +289,7 @@ func main() {
 
 	log.WithFields(log.Fields{"endpoint": *rtSock, "runtime": global.RT.String()}).Info("Container socket connected")
 	if platform == share.PlatformKubernetes {
-		k8sVer, ocVer := global.ORCH.GetVersion()
+		k8sVer, ocVer := global.ORCH.GetVersion(false, false)
 		if flavor == "" && resource.IsRancherFlavor() {
 			flavor = share.FlavorRancher
 			global.ORCH.SetFlavor(flavor)
