+++ share/orchestration/types.go
@@ -22,7 +22,7 @@ type Service struct {
 }
 
 type Driver interface {
-	GetVersion() (string, string)
+	GetVersion(reGetK8sVersion, reGetOcVersion bool) (string, string)
 	SetIPAddrScope(ports map[string][]share.CLUSIPAddr, meta *container.ContainerMeta, nets map[string]*container.Network)
 	GetService(meta *container.ContainerMeta) *Service
 	GetPlatformRole(meta *container.ContainerMeta) (string, bool) // return platform type and if container should be secured
